third_party/blink/web_tests/external/wpt/clipboard-apis/clipboard-item.https-expected.txt

This is a testharness.js-based test.
[FAIL] getType(DOMString valid type) converts DOMString to Blob
  assert_equals: expected (string) "text/plain" but got (undefined) undefined
[FAIL] getType(DOMString invalid type) converts DOMString to Blob
  assert_equals: expected (string) "not a/real type" but got (undefined) undefined
[FAIL] supports(text/uri-list) returns true
  assert_equals: expected true but got false
Harness: the test ran to completion.
